- ad: add form detail  ,add address number

+++ app/src/main/java/com/android/takeoutapp/bean/UserBean.kt
@@ -8,7 +8,7 @@ class UserBean(
     var id: Int,
     var roomId: Int? = null,
     var roomName: String? = null,
-    var number: Int? = null,
+    var number: Long? = null,
     var address: String? = null
 )
